% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dcm2nii.R
\name{check_dcm2nii}
\alias{check_dcm2nii}
\title{Check Output from \code{dcm2nii}}
\usage{
check_dcm2nii(
  dcm2nii_output,
  include_json = TRUE,
  ignore_roi_if_multiple = FALSE
)
}
\arguments{
\item{dcm2nii_output}{Output from \code{\link{dcm2nii}} command, must have
\code{nii_after} element}

\item{include_json}{Include JSON file if in the output?}

\item{ignore_roi_if_multiple}{If multiple nifti files are found, ignore 
any that have the format if `ROI1.nii`, `ROI2.nii`, etc.  
See \url{https://github.com/rordenlab/dcm2niix/issues/832}}
}
\value{
Character vector of unique nifti filenames
}
\description{
This function checks whether the output from \code{\link{dcm2nii}}
runs correctly in that only one \code{nifti} file is returned if only one exists
}
