% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcm2nii_base.R
\name{dcm2nii}
\alias{dcm2nii}
\title{DICOM 2 NIFTI Converter}
\usage{
dcm2nii(basedir, progdir = system.file(package = "dcm2niir"),
  verbose = TRUE, dcm2niicmd = c("dcm2nii_2009", "dcm2nii", "dcm2niix"))
}
\arguments{
\item{basedir}{(character) directory to run \code{dcm2nii}}

\item{progdir}{(character) directory of bash scripts, no user input needed
unless necessary}

\item{verbose}{(logical) print diagnostic printouts}

\item{dcm2niicmd}{(character) either "dcm2nii" or "dcm2nii_2009", which
are different versions of dcm2nii.  Can also be "dcm2niix" - experimental
and only Mac OSX}
}
\value{
List of result of \code{system} run and indication of an error
}
\description{
Uses Chris Rorden's dcm2nii from
http://www.mccauslandcenter.sc.edu/mricro/mricron/dcm2nii.html to convert
DICOM files to NIfTI format.  Need
to run \code{\link{install_dcm2nii}} before running.
}
\examples{
\dontrun{
### This code will install the dcm2nii functions

dcm2nii_files = system.file("dcm2nii", package= "dcm2niir")

if (!file.exists(dcm2nii_files)){
  install_dcm2nii()
}
}
}

