% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm2nii_bin.R
\name{dcm2nii_bin}
\alias{dcm2nii_bin}
\title{Make path for dcm2nii binary}
\usage{
dcm2nii_bin(progdir = system.file(package = "dcm2niir"),
  dcm2niicmd = c("dcm2niix", "dcm2nii_2009", "dcm2nii"))
}
\arguments{
\item{progdir}{(character) directory of bash scripts, no user input needed unless
binaries were installed elsewhere}

\item{dcm2niicmd}{(character) either "dcm2niix", "dcm2nii", or "dcm2nii_2009", which 
are different versions of dcm2nii.}

\item{verbose}{(logical) print diagnostic printouts}
}
\value{
Character vector
}
\description{
Wrapper for checking if dcm2nii binaries are installed and 
creating a file path to the binary requested.
}
\examples{
\dontrun{
dcm2nii_bin()
}
}

