% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm2nii_version.R
\name{dcm2nii_version}
\alias{dcm2nii_version}
\title{dcm2nii version}
\usage{
dcm2nii_version(..., short = FALSE)
}
\arguments{
\item{...}{Arguments passed to \code{\link{dcm2nii_bin}}.}

\item{short}{report back a short version.}
}
\value{
A character string of the version
}
\description{
Finds the version of the dcm2nii function called
}
\examples{
\dontrun{
dcm2nii_version()
}
}
