% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dcm2nii.R
\name{install_dcm2nii}
\alias{install_dcm2nii}
\title{Install dcm2nii tools}
\usage{
install_dcm2nii(lib.loc = NULL)
}
\arguments{
\item{lib.loc}{a character vector with path names of R libraries. 
Passed to \code{\link{system.file}}}
}
\description{
Install dcm2nii to dcm2niir for conversion tools
}
\examples{
install_dcm2nii()
}

