% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm2nii.R
\name{dcm2nii}
\alias{dcm2nii}
\title{Simple run of DICOM 2 NIFTI Converter}
\usage{
dcm2nii(basedir, progdir = system.file(package = "dcm2niir"),
  verbose = TRUE, dcm2niicmd = c("dcm2niix", "dcm2nii_2009", "dcm2nii"),
  ...)
}
\arguments{
\item{basedir}{(character) directory to get files}

\item{progdir}{(character) directory of bash scripts, no user input needed
unless necessary}

\item{verbose}{(logical) print diagnostic printouts}

\item{dcm2niicmd}{(character) either "dcm2niix", "dcm2nii", or "dcm2nii_2009", which 
are different versions of dcm2nii.}

\item{...}{arguments to be passed to \code{\link{system}}}
}
\value{
Result of \code{system} command.
}
\description{
Uses Chris Rorden's dcm2nii from 
http://www.mccauslandcenter.sc.edu/mricro/mricron/dcm2nii.html to convert
DICOM files to NIfTI format.  Should Need run \code{\link{install_dcm2nii}} before running.
}

