% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dcm2nii.R
\name{install_dcm2nii}
\alias{install_dcm2nii}
\alias{install_dcm2nii_bin}
\title{Install dcm2nii tools}
\usage{
install_dcm2nii(
  lib.loc = NULL,
  overwrite = FALSE,
  from_source = FALSE,
  jpeg = FALSE,
  progdir = NULL,
  cmake_opts = "",
  verbose = TRUE,
  source_clone_dir = NULL,
  git_url = "https://github.com/rordenlab/dcm2niix"
)

install_dcm2nii_bin(
  progdir = system.file(package = "dcm2niir"),
  dcm2niicmd = c("dcm2niix", "dcm2niix_feb2024", "dcm2nii_2009", "dcm2nii"),
  overwrite = FALSE
)
}
\arguments{
\item{lib.loc}{a character vector with path names of R libraries.
Passed to \code{\link{system.file}}}

\item{overwrite}{Force reinstallation if one already present}

\item{from_source}{if this is TRUE, then \code{git} and \code{cmake}
are required and the current version is cloned from 
\url{https://github.com/rordenlab/dcm2niix}}

\item{jpeg}{install using JPEG utilities}

\item{progdir}{Installation directory for executable built}

\item{cmake_opts}{Additional options to pass to \code{cmake}, like 
\code{-DZLIB_IMPLEMENTATION=Cloudflare}}

\item{verbose}{print diagnostic messages}

\item{source_clone_dir}{experimental, where the file should be cloned if 
source install.  Do not use unless you know what you're doing.}

\item{git_url}{URL of \code{git} repository used for building from source}

\item{dcm2niicmd}{(character) either "dcm2niix", "dcm2nii", or "dcm2nii_2009", which 
are different versions of dcm2nii.}
}
\description{
Install dcm2nii to dcm2niir for conversion tools
}
\examples{
install_dir = tempdir()
install_dcm2nii(progdir = install_dir)
}
