% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dcm2nii.R
\name{install_dcm2nii}
\alias{install_dcm2nii}
\title{Install dcm2nii tools}
\usage{
install_dcm2nii(lib.loc = NULL, overwrite = FALSE,
  from_source = FALSE, jpeg = FALSE, progdir = NULL,
  cmake_opts = "")
}
\arguments{
\item{lib.loc}{a character vector with path names of R libraries.
Passed to \code{\link{system.file}}}

\item{overwrite}{Force reinstallation if one already present}

\item{from_source}{if this is TRUE, then \code{git} and \code{cmake}
are required and the current version is cloned from 
\url{https://github.com/rordenlab/dcm2niix}}

\item{jpeg}{install using JPEG utilities}

\item{progdir}{Installation directory for executable built}

\item{cmake_opts}{Additional options to pass to \code{cmake}, like 
\code{-DZLIB_IMPLEMENTATION=Cloudflare}}
}
\description{
Install dcm2nii to dcm2niir for conversion tools
}
\examples{
install_dir = tempdir()
install_dcm2nii(progdir = install_dir)
}
